refactor(git): extract PullResult/PushResult + reason constants to types.py (#577 PR1)
